requirements.txt.txt
streamlit
network
itertools
matplotlib.pyplot